Drasil.System

Description

Define types and functions related to creating a system information database.

System

Types

data System where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: (CommonIdea a, Idea a, Quantity e, Eq e, MayHaveUnit e, Concept e, Quantity h, MayHaveUnit h, Concept h, Quantity i, MayHaveUnit i, Concept i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: SystemKind

, _authors :: People

, _purpose :: Purpose

, _background :: Background

, _scope :: Scope

, _motivation :: Motivation

, _quants :: [e]

, _theoryModels :: [TheoryModel]

, _genDefns :: [GenDefn]

, _dataDefns :: [DataDefinition]

, _instModels :: [InstanceModel]

, _configFiles :: [String]

, _inputs :: [h]

, _outputs :: [i]

, _constraints :: [j]

, _constants :: [ConstQDef]

, _systemdb :: ChunkDB

} -> System

data SystemKind

Constructors

Specification

RunnableSoftware

Notebook

Website

Lenses

class HasSystem c where

Minimal complete definition

system

Methods

system :: Lens' c System

authors :: Lens' c People

background :: Lens' c Background

configFiles :: Lens' c [String]

constants :: Lens' c [ConstQDef]

dataDefns :: Lens' c [DataDefinition]

genDefns :: Lens' c [GenDefn]

instModels :: Lens' c [InstanceModel]

kind :: Lens' c SystemKind

motivation :: Lens' c Motivation

purpose :: Lens' c Purpose

scope :: Lens' c Scope

systemdb :: Lens' c ChunkDB

theoryModels :: Lens' c [TheoryModel]

Functions

whatsTheBigIdea :: System -> IdeaDict

mkSystem :: (CommonIdea a, Idea a, Quantity e, Eq e, MayHaveUnit e, Concept e, Quantity h, MayHaveUnit h, Concept h, Quantity i, MayHaveUnit i, Concept i, HasUID j, Constrained j) => a -> SystemKind -> People -> Purpose -> Background -> Scope -> Motivation -> [e] -> [TheoryModel] -> [GenDefn] -> [DataDefinition] -> [InstanceModel] -> [String] -> [h] -> [i] -> [j] -> [ConstQDef] -> ChunkDB -> System

Reference Database

Types

type Purpose = [Sentence]

Project Example purpose.

type Background = [Sentence]

Project Example background information, used in the What section of README.

type Scope = [Sentence]

Project Example scope.

type Motivation = [Sentence]

Project Example motivation.